# backend/docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "IRMS 后端 API（Gin + Gorm + Swagger）",
        "title": "IRMS API",
        "contact": {},
        "version": "0.1"
    },
    "basePath": "/api",
    "paths": {
...
        "/host-credentials": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "credentials"
                ],
                "summary": "列出主机凭据",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page (\u003e=1); invalid -\u003e 400",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size (1~100); invalid -\u003e 400",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Host ID; invalid -\u003e 400",
                        "name": "host_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtoresp.CredentialListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "credentials"
                ],
                "summary": "创建主机凭据",
                "parameters": [
                    {
                        "description": "HostCredentialCreateRequest",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.HostCredentialCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtoresp.IDResponse"
                        }
                    }
                }
            }
        },
        "/host-credentials/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "credentials"
                ],
                "summary": "更新主机凭据",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Credential ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "HostCredentialUpdateRequest",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.HostCredentialUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtoresp.EmptyResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "credentials"
                ],
                "summary": "删除主机凭据",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Credential ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtoresp.EmptyResponse"
                        }
                    }
                }
            }
        },
        "/host-credentials/{id}/reveal": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "credentials"
                ],
                "summary": "查看主机凭据明文",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Credential ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtoresp.CredentialRevealResponse"
                        }
                    }
                }
            }
        },
...
        "/service-credentials": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "credentials"
                ],
                "summary": "列出服务凭据",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page (\u003e=1); invalid -\u003e 400",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size (1~100); invalid -\u003e 400",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Service ID; invalid -\u003e 400",
                        "name": "service_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtoresp.CredentialListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "credentials"
                ],
                "summary": "创建服务凭据",
                "parameters": [
                    {
                        "description": "ServiceCredentialCreateRequest",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ServiceCredentialCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtoresp.IDResponse"
                        }
                    }
                }
            }
        },
        "/service-credentials/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "credentials"
                ],
                "summary": "更新服务凭据",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Credential ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "ServiceCredentialUpdateRequest",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ServiceCredentialUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtoresp.EmptyResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "credentials"
                ],
                "summary": "删除服务凭据",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Credential ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtoresp.EmptyResponse"
                        }
                    }
                }
            }
        },
        "/service-credentials/{id}/reveal": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "credentials"
                ],
                "summary": "查看服务凭据明文",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Credential ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtoresp.CredentialRevealResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controller.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {},
                "details": {},
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
...
        "dtoresp.CredentialListData": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/vo.CredentialVO"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/dtoresp.Pagination"
                }
            }
        },
        "dtoresp.CredentialListResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {
                    "$ref": "#/definitions/dtoresp.CredentialListData"
                },
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
        "dtoresp.CredentialRevealResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {
                    "$ref": "#/definitions/vo.CredentialRevealVO"
                },
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
        "dtoresp.EmptyData": {
            "type": "object"
        },
        "dtoresp.EmptyResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {
                    "$ref": "#/definitions/dtoresp.EmptyData"
                },
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
...
        "dtoresp.IDData": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "dtoresp.IDResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {
                    "$ref": "#/definitions/dtoresp.IDData"
                },
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
...
        "dtoresp.Pagination": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "request.HostCredentialCreateRequest": {
            "type": "object",
            "required": [
                "account_name",
                "credential_kind",
                "credential_name",
                "host_id"
            ],
            "properties": {
                "account_name": {
                    "type": "string"
                },
                "certificate_pem": {
                    "type": "string"
                },
                "credential_kind": {
                    "type": "string"
                },
                "credential_name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "host_id": {
                    "type": "integer"
                },
                "passphrase": {
                    "type": "string"
                },
                "private_key_pem": {
                    "type": "string"
                },
                "secret": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "request.HostCredentialUpdateRequest": {
            "type": "object",
            "required": [
                "account_name",
                "credential_kind",
                "credential_name",
                "host_id"
            ],
            "properties": {
                "account_name": {
                    "type": "string"
                },
                "certificate_pem": {
                    "type": "string"
                },
                "credential_kind": {
                    "type": "string"
                },
                "credential_name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "host_id": {
                    "type": "integer"
                },
                "passphrase": {
                    "type": "string"
                },
                "private_key_pem": {
                    "type": "string"
                },
                "secret": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "request.ServiceCredentialCreateRequest": {
            "type": "object",
            "required": [
                "account_name",
                "credential_kind",
                "credential_name",
                "service_id"
            ],
            "properties": {
                "account_name": {
                    "type": "string"
                },
                "certificate_pem": {
                    "type": "string"
                },
                "credential_kind": {
                    "type": "string"
                },
                "credential_name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "passphrase": {
                    "type": "string"
                },
                "private_key_pem": {
                    "type": "string"
                },
                "secret": {
                    "type": "string"
                },
                "service_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "request.ServiceCredentialUpdateRequest": {
            "type": "object",
            "required": [
                "account_name",
                "credential_kind",
                "credential_name",
                "service_id"
            ],
            "properties": {
                "account_name": {
                    "type": "string"
                },
                "certificate_pem": {
                    "type": "string"
                },
                "credential_kind": {
                    "type": "string"
                },
                "credential_name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "passphrase": {
                    "type": "string"
                },
                "private_key_pem": {
                    "type": "string"
                },
                "secret": {
                    "type": "string"
                },
                "service_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "vo.CredentialRevealVO": {
            "type": "object",
            "properties": {
                "certificate_pem": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "passphrase": {
                    "type": "string"
                },
                "private_key_pem": {
                    "type": "string"
                },
                "secret": {
                    "type": "string"
                }
            }
        },
        "vo.CredentialVO": {
            "type": "object",
            "properties": {
                "account_name": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "credential_kind": {
                    "type": "string"
                },
                "credential_name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "host_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "service_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}